requirements.txt
streamlit==1.35.0
paddleocr==2.7.3
paddlepaddle==3.2.2
opencv-python==4.6.0.66
PyMuPDF==1.24.3
numpy==1.26.4
Pillow==10.2.0
pandas==2.2.0
python-dateutil==2.9.0
opencv-python-headless<4.10
pytesseract
pdf2image 
fitz 
PyMuPDF 
pdfplumber
python-docx
fuzzywuzzy[speedup] 
python-Levenshtein
numpy<2.0.0
llama_cpp_python
huggingface_hub 
sentence_transformers
plotly
torch 
torchvision 
torchaudio
